[Qwen Code] Декомпозиция bot.py на модули
- utils.py — утилиты (get_weekday_russian, функции даты/времени)
- keyboards.py — фабрики inline-клавиатур
- handlers.py — обработчики команд и callback query
- bot.py — сокращён с 1686 до ~400 строк (ядро + запуск)

Бот протестирован, запущен, ошибок в логах нет.
